chore: abstract login/signup
